[DF-7335] Win 10 warnings fix.
Former-commit-id: ded227ee8bfed5e2dc5ba16f80992b57c8ca6315

+++ Sources/Internal/Concurrency/ThreadWin32.cpp
@@ -72,7 +72,7 @@ void Thread::Start()
     auto hdl = _beginthreadex
         (
         0, // Security attributes
-        stackSize,
+        static_cast<DWORD>(stackSize),
         ThreadFunc,
         this,
         0,
